+++ govcd/openapi_test.go
@@ -8,7 +8,7 @@ package govcd
 
 import (
 	"encoding/json"
-	"fmt"
+	"net/http"
 	"net/url"
 	"regexp"
 	"strings"
@@ -19,24 +19,29 @@ import (
 	. "gopkg.in/check.v1"
 )
 
-// Test_OpenApiRawJsonAudiTrail uses low level GET function to test out that pagination really works. It is an example
+// Test_OpenApiRawJsonAuditTrail uses low level GET function to test out that pagination really works. It is an example
 // how to fetch response from multiple pages in RAW json messages without having defined as struct.
-func (vcd *TestVCD) Test_OpenApiRawJsonAudiTrail(check *C) {
-	minimumRequiredApiVersion := "33.0"
-	skipOpenApiEndpointTest(vcd, check, "1.0.0/auditTrail", minimumRequiredApiVersion)
-
-	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint("1.0.0/auditTrail")
+func (vcd *TestVCD) Test_OpenApiRawJsonAuditTrail(check *C) {
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointAuditTrail
+	skipOpenApiEndpointTest(vcd, check, endpoint)
+	apiVersion, err := vcd.client.Client.checkOpenApiEndpointCompatibility(endpoint)
 	check.Assert(err, IsNil)
+
+	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint)
+	check.Assert(err, IsNil)
+
+	// Get a timestamp after which endpoint contains at least 10 elements
+	filterTimeStamp := getAuditTrailTimestampWithElements(10, check, vcd, apiVersion, urlRef)
 
 	// Limit search of audits trails to the last 12 hours so that it doesn't take too long and set pageSize to be 1 result
 	// to force following pages
 	queryParams := url.Values{}
-	filterTime := time.Now().Add(-12 * time.Hour).Format(types.FiqlQueryTimestampFormat)
-	queryParams.Add("filter", "timestamp=gt="+filterTime)
-	queryParams.Add("pageSize", "1")
+	queryParams.Add("filter", "timestamp=gt="+filterTimeStamp)
+	queryParams.Add("pageSize", "1") // pageSize=1 to enforce internal pagination
+	queryParams.Add("sortDesc", "timestamp")
 
 	allResponses := []json.RawMessage{{}}
-	err = vcd.vdc.client.OpenApiGetAllItems(minimumRequiredApiVersion, urlRef, queryParams, &allResponses)
+	err = vcd.vdc.client.OpenApiGetAllItems(apiVersion, urlRef, queryParams, &allResponses)
 
 	check.Assert(err, IsNil)
 	check.Assert(len(allResponses) > 1, Equals, true)
@@ -50,17 +55,19 @@ func (vcd *TestVCD) Test_OpenApiRawJsonAudiTrail(check *C) {
 	check.Assert(len(matches), Equals, len(allResponses))
 }
 
-// Test_OpenApiInlineStructAudiTrail uses low level GET function to test out that get function can unmarshal directly
+// Test_OpenApiInlineStructAuditTrail uses low level GET function to test out that get function can unmarshal directly
 // to user defined inline type
-func (vcd *TestVCD) Test_OpenApiInlineStructAudiTrail(check *C) {
-	minimumRequiredApiVersion := "33.0"
-	skipOpenApiEndpointTest(vcd, check, "1.0.0/auditTrail", minimumRequiredApiVersion)
+func (vcd *TestVCD) Test_OpenApiInlineStructAuditTrail(check *C) {
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointAuditTrail
+	skipOpenApiEndpointTest(vcd, check, endpoint)
+	apiVersion, err := vcd.client.Client.checkOpenApiEndpointCompatibility(endpoint)
+	check.Assert(err, IsNil)
 
-	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint("1.0.0/auditTrail")
+	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint)
 	check.Assert(err, IsNil)
 
 	// Inline type
-	type AudiTrail struct {
+	type AuditTrail struct {
 		EventID      string `json:"eventId"`
 		Description  string `json:"description"`
 		OperatingOrg struct {
@@ -91,14 +98,14 @@ func (vcd *TestVCD) Test_OpenApiInlineStructAudiTrail(check *C) {
 		} `json:"additionalProperties"`
 	}
 
-	allResponses := []*AudiTrail{{}}
+	allResponses := []*AuditTrail{{}}
 
-	// Define FIQL query to find events for the last 24 hours
+	// Define FIQL query to find events for the last 6 hours. At least login operations will already be here on test run
 	queryParams := url.Values{}
-	filterTime := time.Now().Add(-24 * time.Hour).Format(types.FiqlQueryTimestampFormat)
+	filterTime := time.Now().Add(-6 * time.Hour).Format(types.FiqlQueryTimestampFormat)
 	queryParams.Add("filter", "timestamp=gt="+filterTime)
 
-	err = vcd.vdc.client.OpenApiGetAllItems(minimumRequiredApiVersion, urlRef, queryParams, &allResponses)
+	err = vcd.vdc.client.OpenApiGetAllItems(apiVersion, urlRef, queryParams, &allResponses)
 
 	check.Assert(err, IsNil)
 	check.Assert(len(allResponses) > 1, Equals, true)
@@ -128,9 +135,9 @@ func (vcd *TestVCD) Test_OpenApiInlineStructAudiTrail(check *C) {
 // 9. Delete role once again
 func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointRoles
-	minimumRequiredApiVersion, err := vcd.client.Client.checkOpenApiEndpointCompatibility(endpoint)
+	apiVersion, err := vcd.client.Client.checkOpenApiEndpointCompatibility(endpoint)
 	check.Assert(err, IsNil)
-	skipOpenApiEndpointTest(vcd, check, endpoint, minimumRequiredApiVersion)
+	skipOpenApiEndpointTest(vcd, check, endpoint)
 
 	// Step 1 - Get all roles
 	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint)
@@ -145,7 +152,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	}
 
 	allExistingRoles := []*InlineRoles{{}}
-	err = vcd.vdc.client.OpenApiGetAllItems(minimumRequiredApiVersion, urlRef, nil, &allExistingRoles)
+	err = vcd.vdc.client.OpenApiGetAllItems(apiVersion, urlRef, nil, &allExistingRoles)
 	check.Assert(err, IsNil)
 
 	// Step 2 - Get all roles using query filters
@@ -159,7 +166,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 
 		expectOneRoleResultById := []*InlineRoles{{}}
 
-		err = vcd.vdc.client.OpenApiGetAllItems(minimumRequiredApiVersion, urlRef2, queryParams, &expectOneRoleResultById)
+		err = vcd.vdc.client.OpenApiGetAllItems(apiVersion, urlRef2, queryParams, &expectOneRoleResultById)
 		check.Assert(err, IsNil)
 		check.Assert(len(expectOneRoleResultById) == 1, Equals, true)
 
@@ -168,7 +175,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 		check.Assert(err, IsNil)
 
 		oneRole := &InlineRoles{}
-		err = vcd.vdc.client.OpenApiGetItem(minimumRequiredApiVersion, singleRef, nil, oneRole)
+		err = vcd.vdc.client.OpenApiGetItem(apiVersion, singleRef, nil, oneRole)
 		check.Assert(err, IsNil)
 		check.Assert(oneRole, NotNil)
 
@@ -189,7 +196,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 		ReadOnly:  false,
 	}
 	newRoleResponse := &InlineRoles{}
-	err = vcd.client.Client.OpenApiPostItem(minimumRequiredApiVersion, createUrl, nil, newRole, newRoleResponse)
+	err = vcd.client.Client.OpenApiPostItem(apiVersion, createUrl, nil, newRole, newRoleResponse)
 	check.Assert(err, IsNil)
 
 	// Ensure supplied and created structs differ only by ID
@@ -202,7 +209,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	check.Assert(err, IsNil)
 
 	updatedRoleResponse := &InlineRoles{}
-	err = vcd.client.Client.OpenApiPutItem(minimumRequiredApiVersion, updateUrl, nil, newRoleResponse, updatedRoleResponse)
+	err = vcd.client.Client.OpenApiPutItem(apiVersion, updateUrl, nil, newRoleResponse, updatedRoleResponse)
 	check.Assert(err, IsNil)
 
 	// Ensure supplied and response objects are identical (update worked)
@@ -212,19 +219,19 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	deleteUrlRef, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint, newRoleResponse.ID)
 	check.Assert(err, IsNil)
 
-	err = vcd.client.Client.OpenApiDeleteItem(minimumRequiredApiVersion, deleteUrlRef, nil)
+	err = vcd.client.Client.OpenApiDeleteItem(apiVersion, deleteUrlRef, nil)
 	check.Assert(err, IsNil)
 
 	// Step 6 - try to read deleted role and expect error to contain 'ErrorEntityNotFound'
 	// Read is tricky - it throws an error ACCESS_TO_RESOURCE_IS_FORBIDDEN when the resource with ID does not
 	// exist therefore one cannot know what kind of error occurred.
 	lostRole := &InlineRoles{}
-	err = vcd.client.Client.OpenApiGetItem(minimumRequiredApiVersion, deleteUrlRef, nil, lostRole)
+	err = vcd.client.Client.OpenApiGetItem(apiVersion, deleteUrlRef, nil, lostRole)
 	check.Assert(ContainsNotFound(err), Equals, true)
 
 	// Step 7 - test synchronous POST and PUT functions (because Roles is a synchronous OpenAPI endpoint)
 	newRole.ID = "" // unset ID as it cannot be set for creation
-	err = vcd.client.Client.OpenApiPostItemSync(minimumRequiredApiVersion, createUrl, nil, newRole, newRoleResponse)
+	err = vcd.client.Client.OpenApiPostItemSync(apiVersion, createUrl, nil, newRole, newRoleResponse)
 	check.Assert(err, IsNil)
 
 	// Ensure supplied and created structs differ only by ID
@@ -237,7 +244,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	check.Assert(err, IsNil)
 
 	updatedRoleResponse2 := &InlineRoles{}
-	err = vcd.client.Client.OpenApiPutItem(minimumRequiredApiVersion, updateUrl2, nil, newRoleResponse, updatedRoleResponse2)
+	err = vcd.client.Client.OpenApiPutItem(apiVersion, updateUrl2, nil, newRoleResponse, updatedRoleResponse2)
 	check.Assert(err, IsNil)
 
 	// Ensure supplied and response objects are identical (update worked)
@@ -247,21 +254,46 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	deleteUrlRef2, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint, newRoleResponse.ID)
 	check.Assert(err, IsNil)
 
-	err = vcd.client.Client.OpenApiDeleteItem(minimumRequiredApiVersion, deleteUrlRef2, nil)
+	err = vcd.client.Client.OpenApiDeleteItem(apiVersion, deleteUrlRef2, nil)
 	check.Assert(err, IsNil)
 
 }
 
-// skipOpenApiEndpointTest is a helper to skip tests for particular unsupported OpenAPI endpoints
-func skipOpenApiEndpointTest(vcd *TestVCD, check *C, endpoint, requiredVersion string) {
-	constraint := ">= " + requiredVersion
-	if !vcd.client.Client.APIVCDMaxVersionIs(constraint) {
-		maxSupportedVersion, err := vcd.client.Client.maxSupportedVersion()
-		if err != nil {
-			panic(fmt.Sprintf("Could not get maximum supported version: %s", err))
-		}
-		skipText := fmt.Sprintf("Skipping test because OpenAPI endpoint '%s' must satisfy API version constraint '%s'. Maximum supported version is %s",
-			endpoint, constraint, maxSupportedVersion)
-		check.Skip(skipText)
+// getAuditTrailTimestampWithElements helps to pick good timestamp filter so that it doesn't take long time to retrieve
+// too many items
+func getAuditTrailTimestampWithElements(elementCount int, check *C, vcd *TestVCD, apiVersion string, urlRef *url.URL) string {
+	client := vcd.client.Client
+	qp := url.Values{}
+	qp.Add("pageSize", "128")
+	qp.Add("sortDesc", "timestamp") // Need to get the newest
+	req := client.newOpenApiRequest(apiVersion, qp, http.MethodGet, urlRef, nil)
+
+	resp, err := client.Http.Do(req)
+	check.Assert(err, IsNil)
+
+	type AuditTrailTimestamp struct {
+		Timestamp string `json:"timestamp"`
 	}
+
+	onePageAuditTrail := make([]AuditTrailTimestamp, 1)
+	onePageResponse := &types.OpenApiPages{}
+	err = decodeBody(types.BodyTypeJSON, resp, &onePageResponse)
+	check.Assert(err, IsNil)
+
+	err = resp.Body.Close()
+	check.Assert(err, IsNil)
+
+	err = json.Unmarshal(onePageResponse.Values, &onePageAuditTrail)
+	check.Assert(err, IsNil)
+
+	var singleElement AuditTrailTimestamp
+
+	// Find newest element limited by provided elementCount
+	if len(onePageAuditTrail) < elementCount {
+		singleElement = onePageAuditTrail[(len(onePageAuditTrail) - 1)]
+	} else {
+		singleElement = onePageAuditTrail[(elementCount - 1)]
+	}
+	return singleElement.Timestamp
+
 }