+++ types/v56/constants.go
@@ -123,6 +123,8 @@ const (
 	MimeControlAccess = "application/vnd.vmware.vcloud.controlAccess+xml"
 	// Mime of VM capabilities
 	MimeVmCapabilities = "application/vnd.vmware.vcloud.vmCapabilitiesSection+xml"
+	// Mime of Vdc Compute Policy References
+	MimeVdcComputePolicyReferences = "application/vnd.vmware.vcloud.vdcComputePolicyReferences+xml"
 )
 
 const (
@@ -321,8 +323,13 @@ const (
 
 // These constants allow to construct OpenAPI endpoint paths and avoid strings in code for easy replacement in future.
 const (
-	OpenApiPathVersion1_0_0 = "1.0.0/"
-	OpenApiEndpointRoles    = "roles/"
+	OpenApiPathVersion1_0_0                   = "1.0.0/"
+	OpenApiEndpointRoles                      = "roles/"
+	OpenApiEndpointAuditTrail                 = "auditTrail/"
+	OpenApiEndpointImportableTier0Routers     = "nsxTResources/importableTier0Routers"
+	OpenApiEndpointExternalNetworks           = "externalNetworks/"
+	OpenApiEndpointVdcComputePolicies         = "vdcComputePolicies/"
+	OpenApiEndpointVdcAssignedComputePolicies = "vdcs/%s/computePolicies"
 )
 
 // Header keys to run operations in tenant context
@@ -331,4 +338,15 @@ const (
 	HeaderTenantContext = "X-VMWARE-VCLOUD-TENANT-CONTEXT"
 	// HeaderAuthContext requires the Org name of the tenant
 	HeaderAuthContext = "X-VMWARE-VCLOUD-AUTH-CONTEXT"
+)
+
+const (
+	// ExternalNetworkBackingTypeNsxtTier0Router defines backing type of NSX-T Tier-0 router
+	ExternalNetworkBackingTypeNsxtTier0Router = "NSXT_TIER0"
+	// ExternalNetworkBackingTypeNsxtVrfTier0Router defines backing type of NSX-T Tier-0 VRF router
+	ExternalNetworkBackingTypeNsxtVrfTier0Router = "NSXT_VRF_TIER0"
+	// ExternalNetworkBackingTypeNetwork defines vSwitch portgroup
+	ExternalNetworkBackingTypeNetwork = "NETWORK"
+	// ExternalNetworkBackingDvPortgroup refers distributed switch portgroup
+	ExternalNetworkBackingDvPortgroup = "DV_PORTGROUP"
 )